Refactor `run` command
