+++ crates/subspace-gateway/Cargo.toml
@@ -1,7 +1,10 @@
 [package]
 name = "subspace-gateway"
 version = "0.1.0"
-authors = ["Teor <teor@riseup.net>"]
+authors = [
+    "Teor <teor@riseup.net>",
+    "Shamil Gadelshin <shamilgadelshin@gmail.com>"
+]
 description = "A Subspace Network data gateway."
 edition = "2021"
 license = "MIT OR Apache-2.0"
